# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PocketChange therapy practice management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  practitioner: "practitioner",
  client: "client",
} as const;

export interface RegisterBody {
  email: string;
  name: string;
  password: string;
  role: RegisterBodyRole;
  securityQuestion?: string;
  securityAnswer?: string;
}

export interface LinkClientBody {
  accessCode: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  practitioner: "practitioner",
  client: "client",
} as const;

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  role: AuthUserRole;
  /** @nullable */
  linkedClientId?: number | null;
  /** @nullable */
  accessCode?: string | null;
}

export interface VerifySecurityQuestionBody {
  email: string;
  answer: string;
}

export interface VerifyQuestionResponse {
  resetToken: string;
  question: string;
}

export interface ResetPasswordBody {
  resetToken: string;
  newPassword: string;
}

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  accessCode: string;
  /** @nullable */
  notes?: string | null;
  practitionerId: number;
  createdAt: string;
  /** @nullable */
  goalCount?: number | null;
  /** @nullable */
  lastCheckin?: string | null;
  /** @nullable */
  averageProgress?: number | null;
}

export interface CreateClientBody {
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateClientBody {
  name?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
}

/**
 * @nullable
 */
export type GoalType = (typeof GoalType)[keyof typeof GoalType] | null;

export const GoalType = {
  process: "process",
  outcome: "outcome",
} as const;

export interface Goal {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  type?: GoalType;
  /** @nullable */
  parentGoalId?: number | null;
  clientId: number;
  /** @nullable */
  currentProgress?: number | null;
  /** @nullable */
  targetDate?: string | null;
  createdAt: string;
  subGoals: Goal[];
}

/**
 * @nullable
 */
export type CreateGoalBodyType =
  | (typeof CreateGoalBodyType)[keyof typeof CreateGoalBodyType]
  | null;

export const CreateGoalBodyType = {
  process: "process",
  outcome: "outcome",
} as const;

export interface CreateGoalBody {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  type?: CreateGoalBodyType;
  /** @nullable */
  parentGoalId?: number | null;
  /** @nullable */
  targetDate?: string | null;
}

/**
 * @nullable
 */
export type UpdateGoalBodyType =
  | (typeof UpdateGoalBodyType)[keyof typeof UpdateGoalBodyType]
  | null;

export const UpdateGoalBodyType = {
  process: "process",
  outcome: "outcome",
} as const;

export interface UpdateGoalBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  type?: UpdateGoalBodyType;
  /** @nullable */
  targetDate?: string | null;
}

export interface Checkin {
  id: number;
  goalId: number;
  percentage: number;
  /** @nullable */
  whatWorked?: string | null;
  /** @nullable */
  whatGotInTheWay?: string | null;
  /** @nullable */
  whatCouldYouDoDifferently?: string | null;
  /** @nullable */
  barrierTags?: string | null;
  createdAt: string;
}

export interface CreateCheckinBody {
  percentage: number;
  /** @nullable */
  whatWorked?: string | null;
  /** @nullable */
  whatGotInTheWay?: string | null;
  /** @nullable */
  whatCouldYouDoDifferently?: string | null;
  /** @nullable */
  barrierTags?: string | null;
}

export interface PortalData {
  client: Client;
  practitionerName: string;
  goals: Goal[];
}

export interface ActivityItem {
  clientName: string;
  goalTitle: string;
  percentage: number;
  createdAt: string;
}

export interface DashboardData {
  totalClients: number;
  activeClients: number;
  totalCheckins: number;
  clients: Client[];
  recentActivity: ActivityItem[];
}

export interface SmsReminder {
  id: number;
  phone: string;
  days: string;
  time: string;
  timezone: string;
  messageTemplate: string;
}

export interface SmsReminderBody {
  phone: string;
  days: string;
  time: string;
  timezone: string;
  messageTemplate: string;
}