# lib/api-zod/src/generated/types/createCheckinBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PocketChange therapy practice management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCheckinBody {
  percentage: number;
  /** @nullable */
  whatWorked?: string | null;
  /** @nullable */
  whatGotInTheWay?: string | null;
  /** @nullable */
  whatCouldYouDoDifferently?: string | null;
  /** @nullable */
  barrierTags?: string | null;
}