# lib/api-zod/src/generated/types/verifySecurityQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PocketChange therapy practice management API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifySecurityQuestionBody {
  email: string;
  answer: string;
}